refactor drawer/nav and search topbar to be at the page level to DRY out all views code
